requirements.txt
streamlit
google-generativeai
os
API
